+++ src/main/java/datacollection/Goals.java
@@ -2,7 +2,7 @@ package datacollection;
 
 public class Goals {
     private long caloriesOut;
-    private long distance;
+    private float distance;
     private long floors;
     private long steps;
